does this work for you? (#1256)
