+++ test/rng_helper.h
@@ -1,0 +1,21 @@
+#pragma once
+
+namespace at {
+struct PhiloxCudaState;
+}
+
+namespace nvfuser {
+
+enum RNGTest_t {
+  Uniform,
+  Normal,
+};
+
+template <typename T>
+void launch_generate_random_numbers_kernel(
+    cudaStream_t stream,
+    T* output,
+    int64_t size,
+    at::PhiloxCudaState philox_args,
+    RNGTest_t rng_test);
+}
